Move and change enums for notifications
