+++ src/components/helpers/types.ts
@@ -1,10 +1,13 @@
 // Props:
-
 export type Align = 'left' | 'right' | 'center';
 
 export type Position = 'top' | 'right' | 'bottom' | 'left';
 
 export type TooltipVariant = 'dark' | 'light' | 'error';
+
+export type NotificationVariant = 'success' | 'info' | 'warning' | 'error';
+
+export type Duration = 'none' | 'short' | 'medium' | 'long';
 
 // Others:
 export type Point = { x: number; y: number };
