+++ src/components/zen-notification/zen-notification.tsx
@@ -1,7 +1,8 @@
 import { Component, Host, h, Prop } from '@stencil/core';
 import { renderIcon } from '../helpers/fa-icons';
 import { faTimes } from '@fortawesome/pro-solid-svg-icons';
-import { getIcon, getTimeout, Duration, NotificationVariant } from '../helpers/helpers';
+import { Duration, NotificationVariant } from '../helpers/types';
+import { getIcon, getTimeout } from './functions';
 
 @Component({
   tag: 'zen-notification',
@@ -12,7 +13,7 @@ export class ZenNotification {
   div: HTMLElement = undefined;
 
   /** Variant  */
-  @Prop() readonly variant: NotificationVariant = NotificationVariant.SUCCESS;
+  @Prop() readonly variant: NotificationVariant = 'success';
 
   /** Height */
   @Prop() readonly height: string = '5rem';
@@ -27,7 +28,7 @@ export class ZenNotification {
   @Prop() readonly message: string = '';
 
   /** Hide duration */
-  @Prop() readonly dismissDuration: Duration = Duration.MEDIUM;
+  @Prop() readonly dismissDuration: Duration = 'medium';
 
   /** Can dismiss */
   @Prop() readonly dismiss: boolean = false;
@@ -37,7 +38,7 @@ export class ZenNotification {
   }
 
   componentDidRender(): void {
-    if (this.dismissDuration !== Duration.NONE) {
+    if (this.dismissDuration !== 'none') {
       setTimeout(() => {
         this.close(this.div);
       }, getTimeout(this.dismissDuration));
